requirements.txt
torch
transformers>=4.51.0
datasets
accelerate
peft
evaluate
numpy
pandas
tqdm
pyyaml
scikit-learn
regex
bitsandbytes
